+++ core/src/main/java/org/seedstack/seed/core/internal/init/KernelManager.java
@@ -12,24 +12,19 @@ import io.nuun.kernel.api.config.KernelConfiguration;
 import io.nuun.kernel.core.NuunCore;
 import io.nuun.kernel.core.internal.scanner.AbstractClasspathScanner;
 import org.reflections.vfs.Vfs;
-import org.seedstack.seed.diagnostic.DiagnosticManager;
-import org.seedstack.seed.core.Seed;
 import org.seedstack.seed.core.SeedRuntime;
 import org.seedstack.seed.core.internal.scan.ClasspathScanHandler;
 import org.seedstack.seed.core.internal.scan.FallbackUrlType;
-import org.seedstack.shed.ClassLoaders;
+import org.seedstack.seed.diagnostic.DiagnosticManager;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import javax.annotation.Nullable;
-import java.io.IOException;
-import java.io.InputStream;
 import java.net.URL;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Scanner;
 import java.util.ServiceLoader;
 
 public class KernelManager {
@@ -68,20 +63,7 @@ public class KernelManager {
     }
 
     public Kernel createKernel(SeedRuntime seedRuntime, @Nullable KernelConfiguration kernelConfiguration, boolean autoStart) {
-        // Startup message
-        StringBuilder startMessage = new StringBuilder("Starting Seed");
-        if (seedRuntime.getVersion() != null) {
-            startMessage.append(" v").append(seedRuntime.getVersion());
-        }
-        LOGGER.info(startMessage.toString());
-
-        // Banner
-        String banner = getBanner();
-        if (banner != null) {
-            System.out.println(banner);
-        }
-
-        // Kernel
+        long startTime = System.currentTimeMillis();
         if (kernelConfiguration == null) {
             kernelConfiguration = NuunCore.newKernelConfiguration();
         }
@@ -89,7 +71,7 @@ public class KernelManager {
         Kernel kernel = createKernel(kernelConfiguration, seedRuntime.getDiagnosticManager());
         if (autoStart) {
             kernel.start();
-            LOGGER.info("Seed started");
+            LOGGER.info("Seed started in {} second(s)", (System.currentTimeMillis() - startTime) / 1000d);
         }
 
         return kernel;
@@ -97,7 +79,6 @@ public class KernelManager {
 
     public void disposeKernel(Kernel kernel) {
         if (kernel != null && kernel.isStarted()) {
-            LOGGER.info("Stopping Seed");
             kernel.stop();
             LOGGER.info("Seed stopped");
         }
@@ -136,21 +117,5 @@ public class KernelManager {
         });
 
         return kernel;
-    }
-
-    private String getBanner() {
-        InputStream bannerStream = ClassLoaders.findMostCompleteClassLoader(Seed.class).getResourceAsStream("banner.txt");
-        if (bannerStream != null) {
-            try {
-                return new Scanner(bannerStream).useDelimiter("\\Z").next();
-            } finally {
-                try {
-                    bannerStream.close();
-                } catch (IOException e) {
-                    // nothing to do
-                }
-            }
-        }
-        return null;
     }
 }